requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-parler==2.3
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
drf-yasg==1.21.8
eventlet==0.38.2
flower==2.0.1
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
openapi-codec==1.3.2
packaging==24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
tornado==6.4.2
twilio==9.3.7
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
